+++ tawsela/database/migrations/2023_10_16_000955_create_reviews_table.php
@@ -12,10 +12,10 @@ class CreateReviewsTable extends Migration {
 		Schema::create('reviews', function(Blueprint $table) {
 			$table->increments('id');
 			$table->timestamps();
-			$table->integer('user_id')->unsigned();
+			$table->BigInteger('user_id')->unsigned();
 			$table->decimal('rate');
 			$table->text('comment');
-			$table->integer('driver_id')->unsigned();
+			$table->BigInteger('driver_id')->unsigned();
 		});
 	}
