+++ tawsela/database/migrations/2023_10_16_000956_create_conversations_table.php
@@ -12,8 +12,8 @@ class CreateConversationsTable extends Migration {
 		Schema::create('conversations', function(Blueprint $table) {
 			$table->increments('id');
 			$table->timestamps();
-			$table->integer('user_id')->unsigned();
-			$table->integer('driver_id')->unsigned();
+			$table->BigInteger('user_id')->unsigned();
+			$table->BigInteger('driver_id')->unsigned();
 		});
 	}
